refactor back link to texts.tsx
